Restructuring the interpreted CPU
Moved all register handling code to Registers.cs.
Moved the base CPU class to CpuBase.cs, and the interface to ICpu.cs.
